+++ R/jpstat-package.R
@@ -4,7 +4,6 @@
 #' @import vctrs
 #' @importFrom dplyr collect
 #' @importFrom pillar obj_sum type_sum
-#' @importFrom lifecycle deprecated
 #'
 #' @name jpstat
 "_PACKAGE"